+++ packages/orq-rc/src/orq_ai_sdk/models/conversationresponse.py
@@ -104,7 +104,7 @@ class ConversationResponse(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kh0kgpb4trfvj9ssjdr5k5vh"
+        "conv_01kh10xsk10zs3gwjfsrpe53rt"
     )
     r"""Unique conversation identifier with `conv_` prefix."""