+++ packages/orq-rc/src/orq_ai_sdk/models/createconversationop.py
@@ -199,7 +199,7 @@ class CreateConversationResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kh0kgpb6m306c6j2bs7trjqk"
+        "conv_01kh10xsk3z5p4jcs3v81a42sx"
     )
     r"""Unique conversation identifier with `conv_` prefix."""