+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH0KGPX4TG9PNS554DVTVDXT"
+        "01KH10XT4NC7F0TTPRH1CKV2XE"
     )
     r"""The unique identifier of the data source"""