+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH0KGPX29GTGA2PTDJCRVZRH"
+        "01KH10XT4M01QRGPWXPZFFTD7P"
     )
     r"""The unique identifier of the data source"""