+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH0KGPX7KZARWE1NBWANXVBM"
+        "01KH10XT4QYNV53Y0E78PZKA5S"
     )
     r"""The unique identifier of the data source"""