+++ packages/orq-rc/RELEASES.md
@@ -5389,3 +5389,13 @@ Based on:
 - [python v4.3.0-rc.38] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.38] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.38 - packages/orq-rc
+
+## 2026-02-09 11:02:34
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.709.1 (2.812.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.39] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.39] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.39 - packages/orq-rc